// api/src/modules/notification/constants/enums/errors.enum.ts
export enum ErrorMessage {
  NOTIFICATION_ID_EMPTY_ERROR = 'Bạn vui lòng chọn [thông báo].',

  ID_NAN_ERROR = 'Giá trị [id] không hợp lệ.',

  NO_CONTENT = 'Không có dữ liệu hiển thị.',

  OPERATOR_NOTIFICATION_ERROR = 'Lưu thông tin [thông báo] thất bại.',
  NOTIFICATION_NOT_FOUND_ERROR = '[Thông báo] không tồn tại (id: %s)',
}